pyproject.toml
[project]
name = "lighthouse-test"
version = "0.1.0"
description = "Project for Lighthouse technical test"
readme = "README.md"
authors = [
    { name = "Leynier Gutiérrez González", email = "leynier41@gmail.com" },
]
requires-python = ">=3.12"

dependencies = [
    "aws-lambda-powertools>=3.6.0",
    "aws-xray-sdk>=2.14.0",
    "boto3>=1.36.23",
    "fastapi>=0.115.8",
    "mangum>=0.19.0",
]

[dependency-groups]
dev = [
    "aws-cdk-lib>=2.179.0",
    "boto3-stubs[all]>=1.36.23",
    "constructs>=10.4.2",
    "pytest>=8.3.4",
    "pytest-dependency>=0.6.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "ruff>=0.9.6",
    "uv-python-lambda>=0.0.7",
]
